made auth actions its own file
